move project search function out of collections class to ProjectCollection class
